crates: move code away from the main module
