+++ src/crates/sources/github.rs
@@ -1,6 +1,9 @@
-use crates::{lists::List, Crate, GitHubRepo};
+use crates::{lists::List, Crate};
+use dirs::GH_MIRRORS_DIR;
 use errors::*;
 use std::borrow::Cow;
+use std::path::{Path, PathBuf};
+use std::str::FromStr;
 
 static CACHED_LIST: &'static str =
     "https://raw.githubusercontent.com/rust-ops/rust-repos/master/data/github.csv";
@@ -59,5 +62,50 @@ impl List for GitHubList {
         }
 
         Ok(list)
+    }
+}
+
+#[derive(Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize, Clone)]
+pub struct GitHubRepo {
+    pub org: String,
+    pub name: String,
+}
+
+impl GitHubRepo {
+    pub(crate) fn slug(&self) -> String {
+        format!("{}/{}", self.org, self.name)
+    }
+
+    pub(crate) fn url(&self) -> String {
+        format!("https://github.com/{}/{}", self.org, self.name)
+    }
+
+    pub(crate) fn mirror_dir(&self) -> PathBuf {
+        GH_MIRRORS_DIR.join(format!("{}.{}", self.org, self.name))
+    }
+
+    pub(in crates) fn prepare(&self, dest: &Path) -> Result<()> {
+        ::git::shallow_clone_or_pull(&self.url(), &self.mirror_dir())?;
+        ::utils::fs::copy_dir(&self.mirror_dir(), &dest)?;
+        Ok(())
+    }
+}
+
+impl FromStr for GitHubRepo {
+    type Err = Error;
+
+    fn from_str(input: &str) -> Result<Self> {
+        let mut components = input.split('/').collect::<Vec<_>>();
+        let name = components.pop();
+        let org = components.pop();
+
+        if let (Some(org), Some(name)) = (org, name) {
+            Ok(GitHubRepo {
+                org: org.to_string(),
+                name: name.to_string(),
+            })
+        } else {
+            bail!("malformed repo url: {}", input);
+        }
     }
 }
