+++ src/ex_prepare.rs
@@ -3,7 +3,6 @@ use crates::Crate;
 use dirs::{EXPERIMENT_DIR, TEST_SOURCE_DIR};
 use errors::*;
 use experiments::Experiment;
-use git;
 use results::WriteResults;
 use run::RunCommand;
 use std::fs;
@@ -17,17 +16,6 @@ fn froml_dir(ex_name: &str) -> PathBuf {
 
 fn froml_path(ex_name: &str, name: &str, vers: &str) -> PathBuf {
     froml_dir(ex_name).join(format!("{}-{}.Cargo.toml", name, vers))
-}
-
-impl Experiment {
-    pub fn fetch_repo_crates(&self) -> Result<()> {
-        for repo in self.crates.iter().filter_map(|krate| krate.github()) {
-            if let Err(e) = git::shallow_clone_or_pull(&repo.url(), &repo.mirror_dir()) {
-                ::utils::report_error(&e);
-            }
-        }
-        Ok(())
-    }
 }
 
 #[cfg_attr(feature = "cargo-clippy", allow(match_ref_pats))]
